pyproject.toml
[tool.poetry]
name = "coopersystem"
version = "0.1.0"
description = ""
authors = ["Marcus Paredes <viniciusparedes@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.3"
django = "3.2"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
djangorestframework-simplejwt = "^5.2.0"
psycopg2 = "^2.9.3"
django-environ = "^0.8.1"
django-rest-swagger = "^2.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
